+++ .gitignore
@@ -1,5 +1,7 @@
 .idea
 .DS_Store
+*.zip
+tmp
 
 # Python
 venv
@@ -10,8 +12,3 @@ __pycache__
 .terraform
 .terraform.*
 terraform.*
-
-terraform-docs
-terraform-docs-*
-
-tmp
