--- cloudrun/data/pubsub_message.json
@@ -1,13 +1,0 @@
-{
-  "message": {
-    "attributes": {
-      "logging.googleapis.com/timestamp": "2021-01-01T01:42:42.123456789Z"
-    },
-    "data": "eyJ0aW1lc3RhbXAiOiAiIiwgInJlc291cmNlIjogeyJsYWJlbHMiOiB7ImpvYl9pZCI6ICIiLCAicHJvamVjdF9pZCI6ICIifX0sICJ0ZXh0UGF5bG9hZCI6ICJKb2IgY29tcGxldGVkIHN1Y2Nlc3NmdWxseS4ifQo=",
-    "messageId": "1234567891234567",
-    "message_id": "1234567891234567",
-    "publishTime": "2021-01-01T01:42:42.42Z",
-    "publish_time": "2021-01-01T01:42:42.42Z"
-  },
-  "subscription": "projects/<project-id>/subscriptions/ai-platform-log"
-}
