--- cloudrun/pyproject.toml
@@ -1,25 +1,0 @@
-[tool.poetry]
-name = "ai-platform-notification"
-version = "0.0.4"
-description = ""
-authors = ["sfujiwara <shuhei.fujiwara@gmail.com>"]
-
-[tool.poetry.dependencies]
-python = "^3.8"
-gunicorn = "^20.0.4"
-google-cloud-logging = "^2.0.2"
-google-cloud-pubsub = "^2.2.0"
-fastapi = "^0.63.0"
-uvicorn = "^0.13.4"
-
-[tool.poetry.dev-dependencies]
-black = "19.10b0"
-jupyter = "1.0.0"
-pytest = "^6.2.1"
-
-[tool.black]
-line-length = 90
-
-[build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
