+++ functions/requirements.txt
@@ -1,0 +1,3 @@
+google-cloud-logging
+google-cloud-pubsub
+pydantic
